+++ CHANGELOG.md
@@ -8,10 +8,12 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [unreleased]
  ### Added
 - add caches for ehr.concept, ehr.territory, ehr.language([#1007](https://github.com/ehrbase/ehrbase/pull/1007))
+- unwrap if match additional double quotes ([#1022](https://github.com/ehrbase/ehrbase/pull/1022))
  ### Changed
 - use caffeine cache instead of ehcache as ehcache has unnecessary blocking([#1007](https://github.com/ehrbase/ehrbase/pull/1007))
 - remove unnecessary DB queries([#1007](https://github.com/ehrbase/ehrbase/pull/1007))
  ### Fixed 
+- maintain a correct history of participations([#1016](https://github.com/ehrbase/ehrbase/pull/1016))
 
 ## [0.23.0]
  ### Added